# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AsOfModel": {
        "properties": {
          "requested": {
            "description": "What the caller asked for; `latest` when unspecified.",
            "title": "Requested",
            "type": "string"
          },
          "resolved": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The concrete vintage actually served.",
            "title": "Resolved"
          }
        },
        "required": [
          "requested",
          "resolved"
        ],
        "title": "AsOfModel",
        "type": "object"
      },
...
      "ConformanceRule": {
        "properties": {
          "applies_to_fields": {
            "description": "Fields the rule governs.",
            "items": {
              "type": "string"
            },
            "title": "Applies To Fields",
            "type": "array"
          },
          "code_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Symbol implementing the rule, for code_ref kinds.",
            "title": "Code Ref"
          },
          "effective_from": {
            "description": "First day the rule applies.",
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last day it applied, if superseded.",
            "title": "Effective To"
          },
          "evidence_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hash of the evidence artifact, if taken.",
            "title": "Evidence Sha256"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Where the upstream evidence lives.",
            "title": "Evidence Url"
          },
          "rationale": {
            "description": "Why this decision was taken, in the author's words.",
            "title": "Rationale",
            "type": "string"
          },
          "rule": {
            "description": "The rule stated in one sentence.",
            "title": "Rule",
            "type": "string"
          },
          "rule_family": {
            "description": "Family the rule versions belong to.",
            "title": "Rule Family",
            "type": "string"
          },
          "rule_id": {
            "description": "Registry id of the rule.",
            "title": "Rule Id",
            "type": "string",
            "x-glasswell-glossary": "gt_conformance_rule"
          },
          "rule_kind": {
            "description": "Executor kind, or code_ref for a policy declaration.",
            "title": "Rule Kind",
            "type": "string"
          },
          "source_id": {
            "description": "Source the rule applies to.",
            "title": "Source Id",
            "type": "string"
          },
          "spec": {
            "additionalProperties": true,
            "description": "Executable specification, served verbatim.",
            "title": "Spec",
            "type": "object"
          },
          "stage": {
            "description": "parse, validate, conform or join.",
            "title": "Stage",
            "type": "string"
          },
          "supersedes_rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Rule version this one replaced.",
            "title": "Supersedes Rule Id"
          }
        },
        "required": [
          "rule_id",
          "rule_family",
          "supersedes_rule_id",
          "source_id",
          "stage",
          "applies_to_fields",
          "rule_kind",
          "spec",
          "rule",
          "rationale",
          "evidence_url",
          "evidence_sha256",
          "effective_from",
          "effective_to",
          "code_ref"
        ],
        "title": "ConformanceRule",
        "type": "object"
      },
...
      "EnvelopeModel_ConformanceRuleDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConformanceRuleDetail"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[ConformanceRuleDetail]",
        "type": "object"
      },
...
      "EnvelopeModel_list_ConformanceRule__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ConformanceRule"
            },
            "title": "Data",
            "type": "array"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[list[ConformanceRule]]",
        "type": "object"
      },
...
      "LinksModel": {
        "additionalProperties": true,
        "properties": {
          "explain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pre-built /v1/explain call for every handle in the body.",
            "title": "Explain"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Next page, or null at the end.",
            "title": "Next"
          },
          "self": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Canonical path of this response.",
            "title": "Self"
          }
        },
        "title": "LinksModel",
        "type": "object"
      },
...
      "MetaModel": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "$ref": "#/components/schemas/AsOfModel",
            "description": "Requested and resolved vintage (SB-04 \u00a72.5)."
          },
          "deprecations": {
            "description": "Deprecations in force for this operation.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Deprecations",
            "type": "array"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "JSON Pointer to glossary term_id (DIR-8).",
            "title": "Labels",
            "type": "object"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page; null at the end.",
            "title": "Next Cursor"
          },
          "request_id": {
            "description": "ULID for this request; echoed in every problem body.",
            "title": "Request Id",
            "type": "string"
          },
          "source_freshness": {
            "additionalProperties": true,
            "description": "Per-source retrieval vintage and state.",
            "title": "Source Freshness",
            "type": "object"
          },
          "warnings": {
            "description": "Structured warnings: code, detail, optional pointer.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "as_of"
        ],
        "title": "MetaModel",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "The wire shape of every failure (RFC 9457).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What went wrong on this request.",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field failures, each with a JSON Pointer.",
            "title": "Errors"
          },
          "instance": {
            "description": "Path of the request that failed.",
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "description": "ULID joining this response to the audit trail.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the error code.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Absolute URI of the error code; resolves at /v1/errors/{code}.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Static owner key for this deployment. /healthz and the frontend are open.",
        "in": "header",
        "name": "X-Glasswell-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Glass-box upstream analytics on public data. Every number this API serves carries a\nderivation handle in band, and `GET /v1/explain` walks that handle back to the\nchecksummed government file it came from.\n\nConventions that hold everywhere: responses are `data` / `meta` / `links`; collections\nare cursor-paginated with no offset parameter; failures are RFC 9457 problem documents\nwhose `type` resolves at `/v1/errors/{code}`; and `as_of` selects knowledge time, with\nthe resolved vintage reported back in `meta.as_of`.\n\nThis deployment is the North Dakota slice: wells, geometry, monthly production, the\nlineage spine, the conformance registry, quarantine and the glossary. Forecasts,\neconomics and other basins are not served.",
    "title": "glasswell",
    "version": "0.1.0",
    "x-glasswell-freeze": {
      "checked_by": "tests/contract/openapi_diff.py",
      "criterion": "S1",
      "frozen_on": "2026-08-20",
      "policy": "Additive change only. A path, operation, parameter, response field or enum value that is published here is published for the life of /v1; removing one, or making an optional request parameter required, is a /v2 event (blueprint \u00a73.6.1).",
      "status": "frozen",
      "surface": "v1"
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/conformance": {
      "get": {
        "description": "Every mapping, unit, datum and validity decision the pipeline applies, each with the rationale for taking it and a URL to the upstream evidence. Newest effective date first. Rules are append-only: a correction is a new rule that supersedes the old one, so a past decision stays readable.",
        "operationId": "list_conformance_rules",
        "parameters": [
          {
            "description": "Opaque cursor from meta.next_cursor. There is no offset parameter.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from meta.next_cursor. There is no offset parameter.",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size, 100 by default, 200 at most (SB-07 \u00a79.5 caps the spine collections lower than the data ones).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Page size, 100 by default, 200 at most (SB-07 \u00a79.5 caps the spine collections lower than the data ones).",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Filter to one source.",
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one source.",
              "title": "Source Id"
            }
          },
          {
            "description": "Filter to one rule kind.",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one rule kind.",
              "title": "Kind"
            }
          },
          {
            "description": "Filter to one rule family.",
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one rule family.",
              "title": "Family"
            }
          },
          {
            "description": "Filter to one pipeline stage.",
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one pipeline stage.",
              "title": "Stage"
            }
          },
          {
            "description": "Rules governing this field.",
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Rules governing this field.",
              "title": "Field"
            }
          },
          {
            "description": "Rules in force on this date.",
            "in": "query",
            "name": "effective_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Rules in force on this date.",
              "title": "Effective At"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_list_ConformanceRule__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "cursor_malformed": {
                    "summary": "Cursor malformed",
                    "value": {
                      "detail": "The cursor did not decode to the four declared fields.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Cursor malformed",
                      "type": "/v1/errors/cursor_malformed"
                    }
                  },
                  "cursor_query_mismatch": {
                    "summary": "Cursor does not match this query",
                    "value": {
                      "detail": "The cursor was minted against a different filter set. Continuing would return a page from a different result set (SB-04 \u00a72.3).",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Cursor does not match this query",
                      "type": "/v1/errors/cursor_query_mismatch"
                    }
                  },
                  "validation_failed": {
                    "summary": "Request validation failed",
                    "value": {
                      "detail": "A parameter failed validation, including declared caps. `errors[]` names each offending pointer. Caps are refused, never silently clamped.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Request validation failed",
                      "type": "/v1/errors/validation_failed"
                    }
                  }
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List conformance rules",
        "tags": [
          "quality"
        ],
        "x-glasswell-dataset": {
          "anchors": [],
          "collection_pointer": "",
          "columns": {
            "default": [
              "/rule_id",
              "/source_id",
              "/rule_kind",
              "/rule_family",
              "/stage",
              "/effective_from",
              "/rationale"
            ],
            "hidden": [],
            "hidden_reason": {},
            "sort": "/effective_from"
          },
          "detail_operation": "get_conformance_rule",
          "facets": [
            "source_id",
            "kind",
            "family",
            "stage",
            "field",
            "effective_at"
          ],
          "group": "kitchen",
          "id": "conformance",
          "intro": "nb_dataset_conformance",
          "order": 21,
          "row_id": [
            "/rule_id"
          ],
          "title": "Conformance rules"
        },
        "x-glasswell-request-example": {
          "path": {},
          "query": {
            "limit": 5
          }
        }
      }
    },
    "/v1/conformance/{rule_id}": {
      "get": {
        "description": "The rule, its executable spec, why it exists and the evidence behind it. Ask for `include=applied_by` to get the reverse index: which derivations cited this rule and how many rows each applied it to. Rules are immutable once written.",
        "operationId": "get_conformance_rule",
        "parameters": [
          {
            "description": "Conformance rule id.",
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "description": "Conformance rule id.",
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "description": "Set to applied_by for the reverse index.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Set to applied_by for the reverse index.",
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_ConformanceRuleDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "not_found": {
                    "summary": "Not found",
                    "value": {
                      "detail": "The resource does not exist, or exists but is not visible to this principal. Both answer the same way so probing reveals nothing.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 404,
                      "title": "Not found",
                      "type": "/v1/errors/not_found"
                    }
                  }
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "validation_failed": {
                    "summary": "Request validation failed",
                    "value": {
                      "detail": "A parameter failed validation, including declared caps. `errors[]` names each offending pointer. Caps are refused, never silently clamped.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Request validation failed",
                      "type": "/v1/errors/validation_failed"
                    }
                  }
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "One conformance rule",
        "tags": [
          "quality"
        ],
        "x-glasswell-request-example": {
          "path": {
            "rule_id": "cr_nd_stream_vocab_1"
          },
          "query": {
            "include": "applied_by"
          }
        }
      }
    },
...
  }
}